cherry-pick 修改本模组 API 注册时间，使其在配置文件之后加载，便于配置文件管控
