+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/init/registry/CommonRegistry.java
@@ -1,7 +1,14 @@
 package com.github.tartaricacid.touhoulittlemaid.init.registry;
 
+import com.github.tartaricacid.touhoulittlemaid.TouhouLittleMaid;
+import com.github.tartaricacid.touhoulittlemaid.block.multiblock.MultiBlockManager;
 import com.github.tartaricacid.touhoulittlemaid.entity.backpack.BackpackManager;
 import com.github.tartaricacid.touhoulittlemaid.entity.info.ServerCustomPackLoader;
+import com.github.tartaricacid.touhoulittlemaid.entity.task.TaskManager;
+import com.github.tartaricacid.touhoulittlemaid.entity.task.meal.MaidMealManager;
+import com.github.tartaricacid.touhoulittlemaid.inventory.chest.ChestManager;
+import com.github.tartaricacid.touhoulittlemaid.item.bauble.BaubleManager;
+import com.github.tartaricacid.touhoulittlemaid.util.AnnotatedInstanceUtil;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -11,6 +18,16 @@ public final class CommonRegistry {
     @SubscribeEvent
     public static void onSetupEvent(FMLCommonSetupEvent event) {
         event.enqueueWork(ServerCustomPackLoader::reloadPacks);
-        event.enqueueWork(BackpackManager::initItemIndex);
+        event.enqueueWork(CommonRegistry::modApiInit);
+    }
+
+    private static void modApiInit() {
+        TouhouLittleMaid.EXTENSIONS = AnnotatedInstanceUtil.getModExtensions();
+        TaskManager.init();
+        BackpackManager.init();
+        BaubleManager.init();
+        MultiBlockManager.init();
+        ChestManager.init();
+        MaidMealManager.init();
     }
 }
